requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
cffi==1.14.6
channels==3.0.4
channels-redis==3.3.0
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
Django==3.2.6
django-crispy-forms==1.12.0
django-filter==2.4.0
django-safedelete==1.0.0
django-simple-history==3.0.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
hiredis==2.0.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
Markdown==3.3.4
msgpack==1.0.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pyOpenSSL==20.0.1
pytz==2021.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.0
zope.interface==5.4.0
